Fix and improvements for Li & Ma image functions.
Before in compute_lima_on_off_image the flux image computation was
broken, a copy & paste error from compute_lima_image with a
variable that didn't exist.

I split it out into a utility function to avoid the copy & paste,
and left a TODO comment / question how to possibly make it better later.

I also renamed the convolved images to `_conv` instead of just `_`.
because I found it very hard to see the one character difference
and know which images are convolved.