Added write functionality to IRF classes #1157
Added a function (to_table) to EffectiveAreaTable2D, EnergyDispersion2D and Background3D to allow (CTA) IRFs
to be writen out to fits files. It converts the stored data to bintableHDU which, in combination with each class
can be saved to one fits file. Still need to implement this for the PSF classes.