Fix random seed and speed up test
Fix the random seed in `test_feldman_cousins.py::test_numerical_confidence_interval_values`, so we know that for this set for random numbers the result reaches the tolerance requested. For an arbitrary random seed, a bigger sample of random numbers would be needed to ensure that the tolerance is met in all cases, which would increase the computation.

Decrease the size of the random number sample to speed up the test.