all-in one spectral model corrections
Add optional log-parabola and exponential cut-off corrections  to TemplateSpectralModel, SkyDiffuseCube and background model.
Allows more flexibility in the models re-optimisation for example it could be useful if they are extrapolated in the fitting range.

Maybe adding a similar all-in one spectral model could be useful for classification and automatise decisions on the optimal shape, no ?