Fix wrong off position in make_theta_squared_table, fixes #3373
If no off position was defined by the user, the off position is
created by the function mirroring the the source position at the
pointing position.

However, due to how the check was performed, this was only done for the
first observation, resulting in wrong off positions for the case of
multiple observations with different pointing positions.

Co-authored-by: Simone Mender <simone.mender@tu-dortmund.de>
Co-authored-by: Lena Linhoff <lena.linhoff@tu-dortmund.de>
Co-authored-by: Axel Donath <axel.donath@mpi-hd.mpg.de>