Change API of ReflectedRegionsFinder
The ReflectedRegionsFinder now takes configuration options
as `__init__` arguments but the `region` and `center` are
now parameters to the `run` method.

This enables creating a `ReflectedRegionsFinder` and calling it
for each region / center in the `ReflectedRegionsBackgroundMaker`
instead of creating new finders over and over.

The run method returns the found regions and their wcs object.

The reliance on caching was removed and the code simplified by that.

Co-authored-by: Axel Donath <axel.donath@mpi-hd.mpg.de>