Apply rad max for individual event energy
Instead of evaluating rad_max at the bin center of the given counts
energy axis, we now use the individual event energy.

By making the WobbleRegionsFinder work properly with PointSkyRegion and
adding a check for overlapping regions in the rad_max case, we can now
perform the classical data reduction for point sky analyses.

The ReflectedRegionsFinder does not work and probably cannot be made to
work with PointSkyRegion, since it uses the extent of the previous
region to place the next. This is obviously not possible for
PointSkyRegion.