Implement stub methods of _repr_html_
This commit provides stub methods of `_repr_html_()` for various
gammapy classes; this results in a (possible) nicer display in Jupyter
notebooks. Implements an initial version of #3921.

The `IPython.display.display()` function can provide nicer, formatted
representations of objects, depending on the context. In Jupyter
notebooks, this is done through an optional `_repr_html_()` method,
which is called (if available) for the result on the last line of a
notebook cell.

This removes the need for `print(some_class_instance)` in notebooks to
provide the nicer string represenation, and simply using just
`some_class_instance` will be enough. If `_repr_html_()` is not
defined, providing a single object on the last line of a notebook cell
would default to `__repr__()`, but with this stub implementation, it
defaults to `__str__()`. Note that if `__str__()` is not defined
explicitly, Python will defer to `__repr__()`, that is,
`str(some_class_instance)` will default to `repr(some_class_instance)`
if `__str__()` is not implemented.

The stub implementation first tries to return the result of a
`to_html()` method, and if that doesn't exist (as of this commit, this
is the case for all classes), it falls back to returning the string
representation of the class instance (provided by `__str__()`, which
is usually nicer than the default representation (given by
`__repr__`).

If an explicit `to_html()` method exists, this is called instead of
the string representation. Making this a separate method (instead of
altering the private `_repr_html_()`), allows usage elsewhere (outside
of notebooks) and clearer API documentation.

The implementation is done only for base classes; derived classes will
fall back to their (grand)parent's `_repr_html_()` method, but still
end up calling their local `to_html()` or `__str__()` methods if at
least one of these is provided.

Several classes do not have a stub `_repr_html_()` implementation:
- classes derived from the `astropy` or `regions` packages (this could
  be implemented for convenience, but for the moment, this is deferred
  to a possible upstream implementation)
- enums (`Enum` does not allow adding single underscore attributes)
- mixin classes (`gammapy.datasets.spectrum.PlotMixin` appears to be
  the only mixin class)
- jsonencoder derived classes (`maps.io.JsonQuantityencoder`)
- constant/singleton-style classes (e.g., `maps.utils.Invalid...`)
- docs utility classes (since these are not for use in notebooks)
- gammapy.scripts helper classes
- classes serving as context manager (`restore_..._status`)
- xml to dict helper classes