Remove call to plt.legend() in ndmap's plot() method
For LC, datasets consist on two axes: time and energy. Making a call to plt.legend() without arguments triggers the following the following message in gammapy:

"No artists with labels found to put in legend.  Note that artists whose label start with an underscore are ignored when legend() is called with no argument.."

If you plot N FluxPoints, the message is printed N times. 

Related issue: #5510