Fix division in error calculations for compute_flux_doubling
When performing doubling timescale calculations, I noticed that the error values were behaving weirdly. I computed the error propagation formula (and checked it on WolframAlpha) and I think there is a mistake in axes_err_1 and axes_err_2, where the square logarithm of the flux ratio should be under the fraction and not above.